requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
beautifulsoup4==4.8.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
dateutils==0.6.6
django-anymail==4.2
django-appconf==1.0.2
django-ckeditor==5.7.1
django-encrypted-model-fields==0.5.8
django-extensions==2.1.0
django-filter==2.0.0
django-measurement==2.4.0
django-multiselectfield==0.1.8
Django==1.11.15
future==0.16.0
graphene-django==2.2.0
graphene==2.1.3
graphql-core==2.1
graphql-relay==0.4.5
idna==2.7
isoweek==1.3.3
jsonfield==2.0.2
mailchimp3==2.0.11
Markdown==2.6.11
measurement==1.8.0
mysqlclient==1.3.13
nameparser==0.5.7
phonenumbers==8.9.10
Pillow==5.2.0
probableparsing==0.0.1
promise==2.1
pyasn1==0.4.4
pycparser==2.19
pycrypto==2.6.1
python-crfsuite==0.9.6
python-dateutil==2.7.3
python-dotenv==0.10.3
python-keyczar==0.716
pytz==2018.5
raven==6.9.0
requests==2.19.1
Rx==1.6.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.11.0
sympy==0.7.6.1
typing==3.6.6
urllib3==1.23
usaddress==0.5.10
validate-email==1.3
